requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
chromedriver-autoinstaller==0.6.4
cryptography==36.0.2
et-xmlfile==1.1.0
h11==0.13.0
idna==3.3
linkhub==1.6.0
MouseInfo==0.1.3
numpy==1.22.3
opencv-python==4.5.5.64
openpyxl==3.0.9
outcome==1.1.0
packaging==24.0
pdfminer.six==20220319
Pillow==9.0.1
popbill==1.60.1
psutil==5.9.8
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.14.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyOpenSSL==22.0.0
PyPDF2==3.0.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pytweening==1.0.4
pywin32==306
selenium==4.17.2
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.9.0
urllib3==1.26.9
wsproto==1.1.0
